// packages/frontend/src/hooks/__generated__/useAuctionsQuery.graphql.ts
export type AuctionStatus = "ACTIVE" | "CANCELLED" | "COMPLETED" | "PENDING" | "%future added value";
export type useAuctionsQuery$variables = {
  cursor?: string | null | undefined;
  limit?: number | null | undefined;
  status?: AuctionStatus | null | undefined;
  userId?: string | null | undefined;
};
export type useAuctionsQuery$data = {
  readonly auctions: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly bidCount: number;
        readonly createdAt: string;
        readonly currentPrice: number;
        readonly description: string | null | undefined;
        readonly endTime: string;
        readonly id: string;
        readonly imageUrl: string;
        readonly reservePrice: number | null | undefined;
        readonly startPrice: number;
        readonly startTime: string;
        readonly status: AuctionStatus;
        readonly title: string;
        readonly userId: string;
      };
    }>;
    readonly pageInfo: {
      readonly endCursor: string | null | undefined;
      readonly hasNextPage: boolean;
    };
  };
};
export type useAuctionsQuery = {
  response: useAuctionsQuery$data;
  variables: useAuctionsQuery$variables;
};